http://bugs.dwscoalition.org/show_bug.cgi?id=114
Initial commit of Dre's importer.  Some things to note:

* does not import comments/entries yet; the code is there but hasn't been
updated to the new modified way of doing things I've implemented.  I will
work on this soon.

* there's no UI yet, so scheduling an actual import is a manual task that
requires someone with access to the command line to actually do it.

* I will put instructions for how to use/test this on the wiki.

The basic flow of the importer is that an import task is created and
inserted into the import_data table.  Individual jobs that make up a task
are inserted into the import_items table.

The scheduler (bin/import-scheduler) is a daemon process that runs and
looks through the available jobs and schedules TheSchwartz jobs for the
ones that need a job.

The worker (bin/worker/content-importer) actually does the tasks of
importing the various things.

There is still a bunch of random things to do (look for FIXME/TODO) as well
as things that I will be documenting on the wiki/bug.

Patch by Andrea Nall <anall@andreanall.com> and Mark Smith <mark@dreamwidth.org>.