http://bugs.dwscoalition.org/show_bug.cgi?id=114
Add notification event for errors and success and status messages, users automatically get these and can't turn them off.  (I figure if they want to use the importer, we're allowed to give them status updates in their inbox.)  The actual text output of the LJ::Event::ImportStatus event is pretty bad right now and needs to be stripped and such, but it works for now.

Patch by mark.