This is experimental.  This small change gives us the ability to render BML pages that are referenced without giving them a BML extension in the URI. In other words, you can visit the page:
    http://www.dreamwidth.org/test

And assuming test.bml exists, that page will get used.  The only time this
breaks down is when you have a directory named "test" and a file named
"test.bml" - it will opt for the file, when the behavior before this patch
would have shown the directory.

However, we only have one situation like that in DW: stats.bml and stats/,
but since stats/ simply redirects to stats.bml, this isn't a problem.

If this patch turns out to cause problems, it will be reverted. :-)

Patch by mark.