This patch adds the ability for us to use jQuery enabled pages.
More generally, this allows the ability for us to define resource groups
for JavaScript (and other) files to be stored in, so we can potentially
offer different versions of the resource files for different parameters.

This behavior is also experimental.

No change is necessary for developers.  If you just ignore this commit,
you can continue working as you always have.  However, if you wish to use
jQuery on your pages (and are not currently using existing LJ libraries -
they're incompatible) then you should look at the included shop.bml file.

Patch by mark.