http://bugs.dwscoalition.org/show_bug.cgi?id=116
This is a checkpoint commit, the shop is non-functional still, it doesn't let
you do much more other than mess around with your shopping cart.

There's a lot of backend functionality for shops and carts and items.  The
very boring pages in this patch show how to work the backend functions.

ALSO, it is hereby acknowledged that some percentage of this is DW-specific
and needs to be moved to a better location and made more generic.  But since
we want the store interface to be dw-free, this is going here first so it
gets the right licensing explicitly.

Patch by mark.