http://bugs.dwscoalition.org/show_bug.cgi?id=936
a) if site=x, with x being the local site is used, the tag will be treated as though it has no site= attribute.

b) if other people use the Dreamwidth code, links to Dreamwidth with site=dreamwidth.org are parsed correctly and get the correct icon.


Patch by yvi.