http://bugs.dwscoalition.org/show_bug.cgi?id=1433
Allow to sort icons by either upload order (current behavior/default) or by keyword.

On your /icons page, the keyword sort results in multiple instances of the icon showing up. Icons with no keywords (but have a nominal keyword of "pic#1234") are sorted to the bottom of the list.

When doing /editicons, the keyword sort sorts only by the first keyword on the list.

Patch by allen.