http://bugs.dwscoalition.org/show_bug.cgi?id=361
Disable quick reply if you are not allowed to comment to this journal (setting + not a member or not on the access list). Previously, you would only find out once you had typed a comment and hit post; now, you will be redirected to the full form with an error message if you click on the reply link. Also some refactoring.

Patch from LiveJournal, with some tweaks by <user name="kareila">.