http://bugs.dwscoalition.org/show_bug.cgi?id=712
Enforce ordering on inclusion of resources, so that we can make site scheme CSS come before per-page CSS. Includes a hack to make some pages respect old behavior (so that colors in per-page CSS don't clash with site scheme colors); this is a temporary measure and should be removed as the CSS framework is made consistent.

Patch by fu.