http://bugs.dwscoalition.org/show_bug.cgi?id=943
Add the ability to apply a minsecurity for each remote account you can crosspost to. Behaves the same way as local minsecurity, and retains the minsecurity on edit (if you set a minsecurity of access-only, if you edit an entry to public here the security will remain friends-only on the other site. If you edit an entry to private here, the security will be changed to private on the other site.).

Patch by allen.