http://bugs.dwscoalition.org/show_bug.cgi?id=2981
If no specific font is set, default to font_base and not just font_fallback (otherwise users have to set the same font over and over for each element they have a setting for). May change fonts for styles with a font_base but without specific fonts.

Patch by ninetydegrees.