http://bugs.dwscoalition.org/show_bug.cgi?id=3165
Don't push the email onto the list if it doesn't exist/we can't show it (to avoid having an undef value in the list). Fix some text message link display issues as well. If the user doesn't have the cap to use a dw.org email, map their cap to the equivalent display option which doesn't display the local email, so we can then vastly simplify the logic of its callers.

Patch by kareila.