http://bugs.dwscoalition.org/show_bug.cgi?id=2351
Have username.dreamwidth.org/archive point to current year, instead of the latest posted year. If there were no entries posted this year, intelligently fall back onto the last year the user posted or (if there were none) the earliest year in the future that the user posted. Also updates the link in the menu to point to /archive instead of calculating the appropriate year for the user on each page lload, as a performance improvement.

Patch by Erin of <user name="teamdestroyer">.