http://bugs.dwscoalition.org/show_bug.cgi?id=3780
Refactor to parse out the URI into components using a module, rather than using a regex on the whole URL, to make the conditionals more readable, and to pave the way for further whitelisting.

Then add www.youtube-nocookie.com to the valid domains to embed from, following the same rules as youtube.com

Patch by fu.