http://bugs.dwscoalition.org/show_bug.cgi?id=2826
Adds a new 'ban_note' user method backed by a new 'bannotes' database table, and adds a new textarea on the banusers page.  If multiple users are banned on the same form submission, the same note will be applied to all of them.

The note is only visible on the ban page. Community admins share a common set of notes.

Patch by kareila.